extend tests
