+++ tests/continuous_integration/test_02_utils.py
@@ -216,52 +216,3 @@ def test_check_input_factors(outdir):
     # accepts a list of files len==1
     output = cmd([in_file])
     assert in_tfs == output
-
-
-def test_get_binding_tfs():
-    cmd = ananse.utils.get_binding_tfs
-    binding = "tests/data/network/binding.h5"
-    ret = cmd(binding)
-    assert len(ret) == 30
-    assert "NFKB2" in ret
-
-
-def test_view_h5():
-    cmd = ananse.utils.view_h5
-    binding = "tests/data/network/binding.h5"
-
-    # regions and TFs
-    regions = cmd(binding, list_regions=True)
-    assert regions.shape == (61786, 1)
-    tfs = cmd(binding, list_tfs=True)
-    assert tfs.shape == (30, 1)
-
-    # all = regions x TFs
-    complete_view = cmd(binding)
-    assert complete_view.shape == (regions.shape[0], tfs.shape[0])
-
-    # various subsets in either format
-    head_wide = cmd(binding, n=10)
-    assert head_wide.shape == (10, 10)
-    head_long = cmd(binding, n=10, fmt="long")
-    assert head_long.shape == (10 * 10, 3)
-
-    subset_wide = cmd(
-        binding,
-        tfs=["NFKB2", "KLF6"],
-        regions=[
-            "chr10:100000171-100000371",
-            "chr10:100001843-100002043",
-        ],
-    )
-    assert subset_wide.shape == (2, 2)
-    subset_long = cmd(
-        binding,
-        tfs=["NFKB2", "KLF6"],
-        regions=[
-            "chr10:100000171-100000371",
-            "chr10:100001843-100002043",
-        ],
-        fmt="long",
-    )
-    assert subset_long.shape == (2 * 2, 3)
